Clarifies the role and scope of classes
Some classes have been renamed to clarify the role:
 - `Constant` became `Value`
 - `MemoryAccess` became `Reference`

Some classes have been created:
 - `Address`
 - `IndexedAddress`

As the scope of `Value` is simply to hold a numeric value, all
static methods that are address related have been moved in the
`Address` class witch is part of the `mc6809` package as it
is CPU specific.

All the parsing logic of indexed addressing has been removed
from `MemoryManagementUnit` to create `IndexedAddress` that is
only responsible to interpret indexed addressing.

Also, the `Disassembler` now overwrite an existing assembly
file instead of appending the disassembled instrunctions at
the end as this was an unexpected behaviour.