The `MMU` concept was replaced by the `Argument` parsing concept
Now, `Argument.next` parses and return a reference, the effective
address, the immediate value or `Argument.NO_ARGUMENT`. The caller
does not need to know what will be the type of argument returned.

This is easier to use, refactor and test. It also simplifies usage
with register based operations by allowing `Mnemonic.registerOr` to
simply take the already parsed/loaded argument in input.

This is easier to use, refactor and test.