Always restore terminal state
When the application exits with an error code, the tty was
kept in `raw -echo` mode.

Now, the tty is restored to its `cooked echo` mode by a shell
trap on `EXIT`.

Note: The terminal has to be in `raw` mode in order to read
keys one by one.  Otherwise, the user has to hit the return
key in order to send all key event to the application. But,
obviously, the interactive  `shell` does not behave as
expected in `raw` mode.