Extract and test logic operations
All logic operations (and, or, xor, com) were moved from `MC6809E` to
the `Logic` class in order to test them and eventually use them as
function references.

The Complement class is now part of the data manipulation package but,
as it is not used by the mnemonics COM, COMA or COMB, it does not
implement the Operation interface and does not modify the CC register.

The Subset class was split to (have an Access (read/write) and an
Output (read only) version in order to use subsets of an output port.