Fixed V, C and H flag test in ADC, ADD SBC, SUB and CMP instructions
Also, those instructions are now part of the Arithmetic class. The
classes Addition and Subtraction are now only used to apply an offset
to a DataOutput. They can't be used to evaluate flags anymore.

Doing so, it is now possible to type and list a BASIC program without
any issue.

Note : Programs using the `SOUND` keyword will hang as the sound
hardware (PIA, DAC, IRQ) is not yet emulated.