Moved all (almost) operations in the `Mnemonic` enum
As most of the CPU operations are defined in static functions,
it was possible to directly define de operation to perform
as part of the mnemonic enumerations.

Only stack related and the two register transfer operations are
kept in the CPU class.  This is temporary and they should be moved
soon.

Also, the performance of the parser was increased as the
instruction are stored in a hash map based on the op code instead
of a list.