Support for more clock frequencies
Calcul of the clock frequency is now an adjustment with the actual
uptime (ns) of the application. This provides better accuracy on
the long run.

Also, the timing of all op codes is now defined and the MPU will
wait the amount of cycle before performing the next operation.

TODO : The amount of cycles of long branch up codes still have to
be dynamically calculated.