Support HSync/VSync interrupts
VDG pins HS (hrorizontal sync) and FS (vertical sync) are connected to CA1,
CB1 of the first PIA (the one connected tot he keyboard). IRQA and IRQB of
that same chip are connected to the IRQ pin of the MPU.

The MC6847 class is now ClockAware and triggers HS every 60 cycles (15kHz)
and FS every 15000 cycles (60 Hz). MC6821Port also support CA1 interrupt
without support for low-to-high and high-to-low trigger.