Allow direct operations on PIA input/output events
It is now possible to register on a PIA port on input or output
operations in order to configure the PIA or perform an operation.
This avoid having to use the clock in order to read the PIA port
state each cycle.

The keyboard adapter now uses this technique to perform the key
scan.