Output sound from the DAC
Sound generated by the DAC is now rendered using `AudioSystem` at
a frequency of 44100Hz.  When the value of the DAC is the same for
more than a second, the audio output is disabled.

The resolution of the DAC is 6 bits, unfortunately, this is not
supported on most systems. The resolution was set to 8 bits causing
the volume to be lower than expected.

On Color Basic, the sound can be tested with
 - `SOUND 10,10` or
 - `PLAY "CCCEEDEDCCCEEDEE"`.