Extract `InputStream` from `Buffer`
`Buffer` does not extend `InputStream` but provides an instance
of `InputStream` when invoking `Buffer::input`.

Also, the `AudioFormat` must be provided to the `Speaker` constructor
and is unsigned only on Linux to remove sound clicks.

TODO: On MacOS, there is a big sound click when starting or stoping
the application.  When PCM Signed is used, there is a bigger sound
click before and after each sound sequence.