Allow tape playback from WAVE files
If present, `playback.wav` is read when the motor is turned on.
On `COLOR BASIC`, this allows the `CLOAD` or `CLOADM` command to
load the next program saved in the audio file.

In order to allow reading files from internet archives like
https://colorcomputerarchive.com/repo/Cassettes, the pitch when
reading has been lowered by 0.9x. The reason is still unknown but
might be related to the fact that r8bemu emulates a clock at
895 kHz instead of 1MHz to be more realistic.