Configure the tape and/or script to run
It is now possible to specify the tape cassette to use when
booting the emulator.  If there is a script with the same name
it is sent to the keyboard.

To automatically load a program named "test", a wave file named
"test.wav" must be present in "cassette/test.wav" and a text file
named "test.bas" with the content "CLOAD" must be present in
"script/test.bas".