Allow switching the tape cassette from the user interface
It is now possible to activate the "cassette" option form the toolbar in
order to select the ".WAV" file to read as a tape cassette.  It is also
possible to "rewind" the actual ".WAV" file in order to restart the
playback from the beginning.

NOTE :  Recording is still saved in `./cassette/recording.wav`

The scripts, ROMs and cassettes are now stored in the user folder under
`${HOME}/.r8bemu/{cassette,script,rom}` and the application is compiled
as `r8bemu.jar`. When compiled, the script `r8bemu` is not required and
the application can be started by `java --jar r8bemu.jar`.

When the rom files are absent, a demo program writes on the screen where
the rom files must be placed in order to use them.