Allow keyboard buffering to be disabled
When typing keys, the key pressed was stored in a buffer that "played"
the key on the keyboard for a fixed amount of time and then cleared the
keyboard for the same amount of time before reading another key from the
buffer.  It has the advantage of sending all the keys at an optimal rate
for the scan routine of Color Basic.

In some cases where the key must be pressed for a specific amount of time
we can't use a buffer. This why the buffer may be disabled from command
line or at runtime via the keyboard action of the toolbar.

Also, the keys are not repeated anymore in both buffered and raw mode.