Restore performance of the terminal display mode
A cache stores the glyph and the color of each character already displayed
in order to avod sending commands ton the terminal as the display is scanned
at 60 fps.

Also, the two joystick driver threads are now interrupted when the terminal
receives [CTRL] + C as all background services are managed by `Threads` class
(implementing `Services`).