Allow packaging the application as a JAR or MacOS application
Now, the configuration folder template is part of the JAR thus avoiding
having to start the application with a shell script. It is also possible
to package R8BEmu as a native application, only tested on MacOS.

The location of rom files moved from `.r8bemu/rom/*.rom` to `.r8bemu/*.rom`.
Changes to the configuration folder have to pe performed manually on previous
installations.