Disable fullscreen mode on MacOS
When in fullscreen mode, if the user asks to choose an audio media
(tape cassette), the complete user interface of MacOS freezes, due to
the file chooser window. On Ubuntu, the fullscreen mode still allow
showing windows on top of the full screen application.

So, as MacOS has the ability to let any window to be fullscreen, there
is no need for this option in our application and it has been disabled
on MacOS.