Fix selected/unselected icons layout
Now, the buffered keyboard button renders as if there was two buttons
that are mutually exclusives. The selection look of the mute/unmute
icon have been applied to the buffered keyboard button. Also, the
distance between the two "icons" was adjusted directly in the "png"
files.

The mute/unmute icon was also redrawn to look better.