Simplified disk emulation
As the goal is only to read disk image, all the emulation of the head
position, and track00 detection have been removed. There now is no
difference between the disk controller and the disk drive.

Doing so, the track register in the FD197x chip is always in sync with
the actual head position.