Select the joystick emulated by the keyboard arrows
When selected, the arrows on the keyboard act as a dpad input for the left, right or
both joysticks. The opposite joystick (or none) is then emulated by the numpad of the
keyboard to allow two players without actual gamepad or joystick devices. This is
actually the only way to play multiplayer on MacOS.