Support GIME graphic modes of the Color Computer 3
The GIME graphic renderer supports all modes with composite and RGB color
palettes, text mode renders the 128 characters but the color attributes are
not supported yet. Manual video page switching was removed as it seams
irrelevant to manually seek video pages on the GIME with 512K of displayable
memory.

Terminal graphic and text mode now uses xterm true color mode supporting the
CMP and RGB palette of the GIME, it also supports 40 and 80 character modes.

The number of sides of a disk is now derived from the size of the image.
Images with more than 40 tracks of 18 sectors are automatically considered as
double sided, allowing bigger disk images.

Few non functional changes:
 - Introduction of the analog line concept, renaming `Bit` for `Discrete` and
   `Data`/`Parallel` for `Binary` in their own packages.
 - The DAC on 6 bit is converted to analog using the NumericRange class that
   is also used to convert other ranges like audio samples and joystick axis.
 - `MC6821Port.port(int mask)` allows working on a subset of the port istead
   of having to deal with the 8 lines at once.