Clock speed implementation and distinct video clock thread
The video clock is set to 15kHz (60Hz refresh rate of 250 video lines) on an
independent `EmulatorContext` thread, thus allowing the main clock to vary in
speed.

A clock frequency divider was introduced in `EmulatorContext` to divide the
frequency of calls on the clock bus.  The main clock is then configured for
1780 kHz divided by 2 on default low speed mode.

Divider is affected by the SAM's MPU rate register:
 - Writing on 0xffd9 sets the divider to 1 and the MPU runs at 1780 kHz.
 - Writing on 0xffd8, sets the divider to 2 and the MPU runs at 890 kHz.